Update remote URL behaviour to allow retrieving remote image with Query String urls
Before this, when the remote URL contains a query string (?photoId=18) the file can't be retrieved on some Apache version.
Now the query string is append to the path and check correctly. These files are always served by the controller because of this specificity but are always accessible and resizable.

Signed-off-by: Stéphane HULARD <s.hulard@chstudio.fr>